16일 작업본 fetch

+++ lib/pages/today/today_take_tile.dart
@@ -8,6 +8,8 @@ import 'package:yaksok_project/components/yaksok_page_route.dart';
 import 'package:yaksok_project/main.dart';
 import 'package:yaksok_project/models/medicine_alarm.dart';
 import 'package:yaksok_project/models/medicine_history.dart';
+import 'package:yaksok_project/pages/add_medicine/add_medicine_page.dart';
+import 'package:yaksok_project/pages/bottomsheet/more_action_bottomsheet.dart';
 import 'package:yaksok_project/pages/bottomsheet/time_setting_bottomsheet.dart';
 import 'package:yaksok_project/pages/today/image_detail_page.dart';
 import 'package:yaksok_project/pages/today/today_page.dart';
@@ -28,7 +30,7 @@ class BeforeTakeTile extends StatelessWidget {
 
     return Row(
         children: [
-          _MedicineImageButton(medicineAlarm: medicineAlarm),
+          MedicineImageButton(imagePath: medicineAlarm.imagePath),
           const SizedBox(width: smallSpace), // 여백
           Expanded(
             child: Column(
@@ -50,18 +52,21 @@ class BeforeTakeTile extends StatelessWidget {
         crossAxisAlignment: WrapCrossAlignment.center,
         children: [
           Text('${medicineAlarm.name},', style: textStyle),
-          TileActionButton(
+          TileActionButton( //'지금' 버튼
             onTap: () { //history에 지금 시간 추가
               historyRepository.addHistory(MedicineHistory( //hive db에 takeDateTime 저장
                   medicineId: medicineAlarm.id
                 , alarmTime: medicineAlarm.alarmTime
-                , takeTime: DateTime.now(),
+                , takeTime: DateTime.now()
+                , imagePath: medicineAlarm.imagePath
+                , name: medicineAlarm.name
+                , medicineKey: medicineAlarm.key,
               ));
             },
             title: '지금',
           ),
           Text('|', style: textStyle),
-          TileActionButton(
+          TileActionButton( //'아까' 버튼
             onTap: () => _onPreviousTake(context),
             title: '아까',
           ),
@@ -83,6 +88,9 @@ class BeforeTakeTile extends StatelessWidget {
         medicineId: medicineAlarm.id
       , alarmTime: medicineAlarm.alarmTime
       , takeTime: takeDateTime
+      , imagePath: medicineAlarm.imagePath
+      , name: medicineAlarm.name
+      , medicineKey: medicineAlarm.key,
       ),
     );
   });
@@ -98,8 +106,8 @@ class AfterTakeTile extends StatelessWidget {
     required this.history,
   }) : super(key: key);
 
-  final MedicineAlarm medicineAlarm;
-  final MedicineHistory history;
+  final MedicineAlarm medicineAlarm; //알람 객체
+  final MedicineHistory history; //복약기록 hive객체
 
   @override
   Widget build(BuildContext context) {
@@ -109,7 +117,7 @@ class AfterTakeTile extends StatelessWidget {
       children: [
         Stack( //이미지 stack (쌓기)
           children: [
-            _MedicineImageButton(medicineAlarm: medicineAlarm),
+            MedicineImageButton(imagePath: medicineAlarm.imagePath),
             CircleAvatar( //이미지 위로 반투명 체크 겹침
               radius: 40,
               backgroundColor: Colors.green.withOpacity(0.8),
@@ -161,7 +169,7 @@ class AfterTakeTile extends StatelessWidget {
     ];
   }
 
-  String get takeTimeStr => DateFormat('HH:mm').format(history.takeTime);
+  String get takeTimeStr => DateFormat('HH:mm').format(history.takeTime); //실 복약시간 date -> string
 
   void _onTap(BuildContext context){
     showModalBottomSheet(
@@ -171,10 +179,10 @@ class AfterTakeTile extends StatelessWidget {
         submitTitle:  '수정',
         bottomWidget: TextButton(
           onPressed: () {
-            historyRepository.deleteHistory(history.key);
+            historyRepository.deleteHistory(history.key); //복약 히스토리 삭제
             Navigator.pop(context);
           },
-          child: Text('약 복용 시간을 되돌리고 싶어요!')
+          child: Text('약 복용 시간을 삭제하고 싶어요!')
         ),
       )
     ).then((takeDateTime){
@@ -187,7 +195,12 @@ class AfterTakeTile extends StatelessWidget {
         history: MedicineHistory(
           medicineId: medicineAlarm.id,
           alarmTime: medicineAlarm.alarmTime,
-          takeTime: takeDateTime
+          takeTime: takeDateTime,
+
+          //추가
+          medicineKey: medicineAlarm.key,
+          imagePath: medicineAlarm.imagePath,
+          name: medicineAlarm.name,
       ),  
       );
     });
@@ -196,10 +209,8 @@ class AfterTakeTile extends StatelessWidget {
   
 
 
-
-
-
-class _MoreButton extends StatelessWidget {
+//더보기 버튼
+class _MoreButton extends StatelessWidget { 
   const _MoreButton({
     Key? key,
     required this.medicineAlarm,
@@ -211,46 +222,84 @@ class _MoreButton extends StatelessWidget {
   Widget build(BuildContext context) {
     return CupertinoButton(
       onPressed: () {
-        medicineRepository.deleteMedicine(medicineAlarm.key);
+        showModalBottomSheet(context: context, builder: (context)=> MoreActionBottomSheet(//context인자로 위젯 반환
+          onPressedUpdate: (){ //약 정보 수정
+            Navigator.push(context, FadePageRoute(page: AddMedicinePage(updateMedicineId:medicineAlarm.id ,))).then((_) => Navigator.maybePop(context));
+          },
+          onPressedDeleteMedicine: (){ //약 정보 삭제
+            //알람 삭제
+            //hive 약 데이터 삭제
+            notification.deleteMultipleAlarm(alarmIds);
+            medicineRepository.deleteMedicine(medicineAlarm.key);
+            Navigator.pop(context);
+          },
+          onPressedDeleteAll: (){ //약 정보와 히스토리 모두삭제
+            //알람 삭제
+            //hive 히스토리 데이터 삭제
+            //hive 약 데이터 삭제
+            notification.deleteMultipleAlarm(alarmIds);
+            historyRepository.deleteAllHistory(keys);
+            medicineRepository.deleteMedicine(medicineAlarm.key);
+            Navigator.pop(context);
+          },
+        )
+        ); 
       },
-      child: const Icon(CupertinoIcons.ellipsis_vertical),
+      child: const Icon(CupertinoIcons.ellipsis_vertical), //점3개 아이콘
     );
+  }
+
+  List<String> get alarmIds{
+    final medicine = medicineRepository.medicineBox.values.singleWhere((element) => element.id == medicineAlarm.id);
+    final alarmIds = medicine.alarms.map((alarmStr) => notification.alarmId(medicineAlarm.id, alarmStr)).toList();
+    return alarmIds;
+  }
+
+  Iterable<int> get keys{
+  final histories = 
+  historyRepository.historyBox.values.where((history) => history.medicineId == medicineAlarm.id && history.medicineKey == medicineAlarm.key);
+  final keys = histories.map((e)=>e.key as int);
+  return keys;
   }
 }
 
-class _MedicineImageButton extends StatelessWidget {
-  const _MedicineImageButton({
+
+//이미지 클릭
+class MedicineImageButton extends StatelessWidget {
+  const MedicineImageButton({
     Key? key,
-    required this.medicineAlarm,
+    required this.imagePath,
   }) : super(key: key);
 
-  final MedicineAlarm medicineAlarm;
+  final String? imagePath;
 
   @override
   Widget build(BuildContext context) {
     return CupertinoButton(
       padding: EdgeInsets.zero, // padding 제거
-      onPressed: medicineAlarm.imagePath == null
+      onPressed: imagePath == null //이미지 클릭
       ? null
       :() {
         Navigator.push(
           context, 
-          FadePageRoute(
-            page: ImageDetailPage(medicineAlarm: medicineAlarm),
+          FadePageRoute( //화면 전환 애니메이션
+            page: ImageDetailPage(imagePath: imagePath!), //이미지 창 크게보기
           ),
         );
       },
       child: CircleAvatar(
         radius: 40,
-        foregroundImage: medicineAlarm.imagePath == null
+        foregroundImage: imagePath == null
         ? null
-        : FileImage(File(medicineAlarm.imagePath!)),
+        : FileImage(File(imagePath!)),
+        child: imagePath == null? Icon(CupertinoIcons.alarm_fill): null,
       ),
     );
   }
 }
 
 
+//지금,아까 타일 버튼
 class TileActionButton extends StatelessWidget {
   const TileActionButton({
     Key? key,
